mcp-servers/ai-cos-mcp/pyproject.toml
[project]
name = "ai-cos-mcp"
version = "0.2.0"
description = "AI Chief of Staff MCP Server + Agent SDK Runners"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "notion-client>=2.2",
    "yt-dlp>=2024.0",
    "youtube-transcript-api>=0.6",
]

[project.optional-dependencies]
agents = [
    "claude-agent-sdk>=0.1",
]
